polish up messages/friends a little

+++ src/endpoints/social/deleteFriendRequest.cr
@@ -13,11 +13,11 @@ CrystalGauntlet.endpoints["/deleteGJFriendRequests20.php"] = ->(context : HTTP::
 
   own_request = params["isSender"]? == "1"
 
-  if DATABASE.scalar("select count(*) from friend_requests where #{own_request ? "to_account_id" : "from_account_id"} = ?", params["targetAccountID"].to_i).as(Int64) == 0
+  if DATABASE.scalar("select count(*) from friend_requests where #{own_request ? "to_account_id" : "from_account_id"} = ? and #{own_request ? "from_account_id" : "to_account_id"}", params["targetAccountID"].to_i, account_id).as(Int64) == 0
     return "-1"
   end
 
-  DATABASE.exec("delete from friend_requests where #{own_request ? "to_account_id" : "from_account_id"} = ?", params["targetAccountID"].to_i)
+  DATABASE.exec("delete from friend_requests where #{own_request ? "to_account_id" : "from_account_id"} = ? and #{own_request ? "from_account_id" : "to_account_id"}", params["targetAccountID"].to_i, account_id)
 
   return "1"
 }
